+++ CHANGELOG.md
@@ -3,8 +3,9 @@
 ## [Unreleased]
 ### Fixed
 - Fixed issue with token retrieving when client is dead
-- Fixed issue with pinging client when client_age < 60
+- Fixed issue with pinging client when `client_age` < 60
 - Fixed issue with key when client created using different key size
+- Fixed compression flag
 
 ## [1.0.0-beta.9]
 ### Fixed